remplacement hello widget par EDT widget
